Add required handlers and define required libraries. (#9)
* Add Additional backend files required to handle
the projects

The current bakcend code is not sufficient to handle all the requests.
Added additional API's to handle more requests.

Signed-off-by: Hruthiksai44 <hruthiksai44@gmail.com>

* Add more handlers and dependencies to backend.

The current backend code does not have enough
handlers and libraries to handle the requests.
Added the required files to make the backend code
fully functional.

Signed-off-by: Hruthiksai44 <hruthiksai44@gmail.com>

* Add more handlers and define required libraries.

The current backend code does not have any required libraries and has
a lack of some handlers.
Added the required handlers to make the backend code functional.

Signed-off-by: Hruthiksai44 <hruthiksai44@gmail.com>

---------

Signed-off-by: Hruthiksai44 <hruthiksai44@gmail.com>
Co-authored-by: Hruthiksai44 <hruthiksai44@gmail.com>