Imports: Fix the language requirements and auth
There are two main things covered in this commit

1. Auth is used as an internal submodule
Which is not the case it is a package.
Therefore when we do go mod tidy in other device it doesn't work

2. The language requirements number expects 1.25+ versions of Go
Which is heavily modern...
We should atleast use the most stable one linux distros provide
And incrementally update them.
FYI Fedora still provides 1.24+ and not 1.25 as that is the most stable
version compatible with it.

Signed-off-by: K Jayatheerth <jayatheerthkulkarni2005@gmail.com>